UNOMI-263: small refactoring
